requirements.txt
anthropic>=0.25.0
requests>=2.31.0
feedparser>=6.0.10
praw>=7.7.1
pyyaml>=6.0.1
pandas>=2.0.0
starknet-py>=0.21.0
python-dotenv>=1.0.0
colorlog>=6.8.0
aiohttp>=3.9.0
asyncio>=3.4.3
